requirements.txt
asgiref
Django
sqlparse
django-bootstrap-v5
djangorestframework